count to gen feature

+++ app/src/main/java/com/euphoiniateam/euphonia/ui/settings/SettingsViewModel.kt
@@ -110,7 +110,7 @@ class SettingsViewModel(
         recordingAudio: Boolean,
         recordingStave: Boolean,
         pianoSize: Float,
-        notesAmount: Float,
+        notesAmount: Int,
         staveOn: Boolean
     ) {
         viewModelScope.launch(Dispatchers.IO) {
